[RENAME] api model명 변경
