+++ src/api/loginapi/SignIn.ts
@@ -1,7 +1,7 @@
 import { http } from '../http';
 import { useMutation } from '@tanstack/react-query';
 import { useInvalidateQueries } from '../../hook/useInvaildQueries';
-import { User } from '../loginmodel';
+import { User } from './model';
 
 
 const MUTATION_KEY = {
